+++ mk_test_env/wizard/__init__.py
@@ -7,4 +7,5 @@
 #
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 #
-from . import wizard
+from . import wizard_mk_test_env
+# from . import wizard_cleanup_test_env